Add functionality to remove multiple games from a user's list

+++ src/main/java/com/checkpoint/modules/library/dto/LibraryDTOs.java
@@ -114,7 +114,7 @@ public class LibraryDTOs {
             Long igdbId,
             String name,
             String coverUrl,
-            Integer releaseYear
+            Long releaseYear
     ) {}
 
     public record GameStatusResponse(
